+++ config/examples/config.distribution-grid-experimental.yaml
@@ -50,9 +50,8 @@ transmission:
   electricity:
     enable: true
     base_network: osm
-
-transmission_projects:
-  enable: false
+    projects:
+      enable: false
 
 data:
   osm:
