+++ doc/configuration.rst
@@ -299,6 +299,32 @@ For non electric carriers, the transmission candidates are built based on
 Delaunay triangulation and optional Gabriel filtering. The minimum degree
 specifies the minimum connectedness of the candidate graph.
 
+.. _transmission_projects_cf:
+
+``electricity.projects``
+------------------------
+
+Allows to define additional transmission projects that will be added to the
+base network, e.g., from the TYNDP 2020 dataset. The projects are read in
+from the CSV files in the subfolder of ``data/transmission_projects/``. New
+transmission projects, e.g. from TYNDP 2024, can be added in a new subfolder
+of transmission projects, e.g. ``data/transmission_projects/tyndp2024`` while
+extending the list of projects in ``transmission.electricity.projects.include``
+in the configuration. The CSV files in the project folder should have the same
+columns as the CSV files in the template folder
+``data/transmission_projects/template``.
+
+.. jsonschema:: ../config/schema.default.json#/$defs/TransmissionProjectsConfig
+   :lift_description:
+   :hide_key: /**/additionalProperties
+
+**YAML Syntax**
+
+.. literalinclude:: ../config/config.default.yaml
+   :language: yaml
+   :start-at:     projects:
+   :end-before:   hydrogen:
+
 .. jsonschema:: ../config/schema.default.json#/$defs/TransmissionConfig
    :lift_description:
    :hide_key: /**/additionalProperties
@@ -446,72 +472,6 @@ overwrite the existing values.
 .. literalinclude:: ../config/config.default.yaml
    :language: yaml
    :start-at:   conventional:
-   :end-before: # docs
-
-.. _lines_cf:
-
-``lines``
-=========
-
-.. jsonschema:: ../config/schema.default.json#/$defs/LinesConfig
-   :lift_description:
-   :hide_key: /**/additionalProperties
-
-**YAML Syntax**
-
-.. literalinclude:: ../config/config.default.yaml
-   :language: yaml
-   :start-at: lines:
-   :end-before: # docs
-
-.. _links_cf:
-
-``links``
-=============
-
-.. jsonschema:: ../config/schema.default.json#/$defs/LinksConfig
-   :lift_description:
-   :hide_key: /**/additionalProperties
-
-**YAML Syntax**
-
-.. literalinclude:: ../config/config.default.yaml
-   :language: yaml
-   :start-at: links:
-   :end-before: # docs
-
-.. _transmission_projects_cf:
-
-``transmission_projects``
-=========================
-
-Allows to define additional transmission projects that will be added to the base network, e.g., from the TYNDP 2020 dataset. The projects are read in from the CSV files in the subfolder of ``data/transmission_projects/``. New transmission projects, e.g. from TYNDP 2024, can be added in a new subfolder of transmission projects, e.g. ``data/transmission_projects/tyndp2024`` while extending the list of ``transmission_projects`` in the ``config.yaml`` by ``tyndp2024``. The CSV files in the project folder should have the same columns as the CSV files in the template folder ``data/transmission_projects/template``.
-
-.. jsonschema:: ../config/schema.default.json#/$defs/TransmissionProjectsConfig
-   :lift_description:
-   :hide_key: /**/additionalProperties
-
-**YAML Syntax**
-
-.. literalinclude:: ../config/config.default.yaml
-   :language: yaml
-   :start-at: transmission_projects:
-   :end-before: # docs
-
-.. _transformers_cf:
-
-``transformers``
-================
-
-.. jsonschema:: ../config/schema.default.json#/$defs/TransformersConfig
-   :lift_description:
-   :hide_key: /**/additionalProperties
-
-**YAML Syntax**
-
-.. literalinclude:: ../config/config.default.yaml
-   :language: yaml
-   :start-at: transformers:
    :end-before: # docs
 
 .. _load_cf:
