+++ rules/build_electricity.smk
@@ -100,9 +100,9 @@ rule base_network:
         countries=config_provider("countries"),
         snapshots=config_provider("snapshots"),
         drop_leap_day=config_provider("enable", "drop_leap_day"),
-        lines=config_provider("lines"),
-        links=config_provider("links"),
-        transformers=config_provider("transformers"),
+        lines=config_provider("transmission", "electricity", "lines"),
+        links=config_provider("transmission", "electricity", "links"),
+        transformers=config_provider("transmission", "electricity", "transformers"),
         clustering=config_provider("clustering", "mode"),
         admin_levels=config_provider("clustering", "administrative"),
     message:
@@ -477,7 +477,10 @@ rule build_line_rating:
     input:
         base_network=resources("networks/base.nc"),
         cutout=lambda w: input_cutout(
-            w, config_provider("lines", "dynamic_line_rating", "cutout")(w)
+            w,
+            config_provider(
+                "transmission", "electricity", "lines", "dynamic_line_rating", "cutout"
+            )(w),
         ),
     output:
         output=resources("dlr.nc"),
@@ -504,9 +507,9 @@ rule build_transmission_projects:
         europe_shape=resources("europe_shape.geojson"),
         transmission_projects=lambda w: [
             "data/transmission_projects/" + name
-            for name, include in config_provider("transmission_projects", "include")(
-                w
-            ).items()
+            for name, include in config_provider(
+                "transmission", "electricity", "projects", "include"
+            )(w).items()
             if include
         ],
     output:
@@ -523,9 +526,11 @@ rule build_transmission_projects:
     resources:
         mem_mb=4000,
     params:
-        transmission_projects=config_provider("transmission_projects"),
-        line_factor=config_provider("lines", "length_factor"),
-        s_max_pu=config_provider("lines", "s_max_pu"),
+        transmission_projects=config_provider("transmission", "electricity", "projects"),
+        line_factor=config_provider(
+            "transmission", "electricity", "lines", "length_factor"
+        ),
+        s_max_pu=config_provider("transmission", "electricity", "lines", "s_max_pu"),
     message:
         "Building transmission projects"
     script:
@@ -537,7 +542,13 @@ rule add_transmission_projects_and_dlr:
         network=resources("networks/base.nc"),
         dlr=lambda w: (
             resources("dlr.nc")
-            if config_provider("lines", "dynamic_line_rating", "activate")(w)
+            if config_provider(
+                "transmission",
+                "electricity",
+                "lines",
+                "dynamic_line_rating",
+                "activate",
+            )(w)
             else []
         ),
         transmission_projects=lambda w: (
@@ -548,7 +559,7 @@ rule add_transmission_projects_and_dlr:
                 resources("transmission_projects/adjust_lines.csv"),
                 resources("transmission_projects/adjust_links.csv"),
             ]
-            if config_provider("transmission_projects", "enable")(w)
+            if config_provider("transmission", "electricity", "projects", "enable")(w)
             else []
         ),
     output:
@@ -561,9 +572,11 @@ rule add_transmission_projects_and_dlr:
     resources:
         mem_mb=4000,
     params:
-        transmission_projects=config_provider("transmission_projects"),
-        dlr=config_provider("lines", "dynamic_line_rating"),
-        s_max_pu=config_provider("lines", "s_max_pu"),
+        transmission_projects=config_provider("transmission", "electricity", "projects"),
+        dlr=config_provider(
+            "transmission", "electricity", "lines", "dynamic_line_rating"
+        ),
+        s_max_pu=config_provider("transmission", "electricity", "lines", "s_max_pu"),
     message:
         "Adding transmission projects and dynamic line ratings"
     script:
@@ -677,8 +690,12 @@ rule simplify_network:
         aggregation_strategies=config_provider(
             "clustering", "aggregation_strategies", default={}
         ),
-        p_max_pu=config_provider("links", "p_max_pu", default=1.0),
-        p_min_pu=config_provider("links", "p_min_pu", default=-1.0),
+        p_max_pu=config_provider(
+            "transmission", "electricity", "links", "p_max_pu", default=1.0
+        ),
+        p_min_pu=config_provider(
+            "transmission", "electricity", "links", "p_min_pu", default=-1.0
+        ),
     message:
         "Simplifying network"
     script:
@@ -753,7 +770,9 @@ rule cluster_network:
             "electricity", "conventional_carriers", default=[]
         ),
         max_hours=config_provider("electricity", "max_hours"),
-        length_factor=config_provider("lines", "length_factor"),
+        length_factor=config_provider(
+            "transmission", "electricity", "lines", "length_factor"
+        ),
         cluster_mode=config_provider("clustering", "mode"),
         copperplate_regions=config_provider("clustering", "copperplate_regions"),
     message:
@@ -817,8 +836,12 @@ rule add_electricity:
     resources:
         mem_mb=10000,
     params:
-        line_length_factor=config_provider("lines", "length_factor"),
-        link_length_factor=config_provider("links", "length_factor"),
+        line_length_factor=config_provider(
+            "transmission", "electricity", "lines", "length_factor"
+        ),
+        link_length_factor=config_provider(
+            "transmission", "electricity", "links", "length_factor"
+        ),
         scaling_factor=config_provider("load", "scaling_factor"),
         countries=config_provider("countries"),
         snapshots=config_provider("snapshots"),
@@ -860,8 +883,8 @@ rule prepare_network:
         mem_mb=4000,
     params:
         time_resolution=config_provider("clustering", "temporal", "resolution_elec"),
-        links=config_provider("links"),
-        lines=config_provider("lines"),
+        links=config_provider("transmission", "electricity", "links"),
+        lines=config_provider("transmission", "electricity", "lines"),
         co2base=config_provider("electricity", "co2base"),
         co2limit_enable=config_provider("electricity", "co2limit_enable", default=False),
         co2limit=config_provider("electricity", "co2limit"),
@@ -960,7 +983,7 @@ rule build_osm_network:
     params:
         countries=config_provider("countries"),
         voltages=config_provider("electricity", "voltages"),
-        line_types=config_provider("lines", "types"),
+        line_types=config_provider("transmission", "electricity", "lines", "types"),
         under_construction=config_provider("osm_network_release", "under_construction"),
         remove_after=config_provider("osm_network_release", "remove_after"),
     message:
