+++ rules/build_sector.smk
@@ -1766,9 +1766,11 @@ rule prepare_sector_network:
         sector=config_provider("sector"),
         industry=config_provider("industry"),
         renewable=config_provider("renewable"),
-        lines=config_provider("lines"),
+        lines=config_provider("transmission", "electricity", "lines"),
         pypsa_eur=config_provider("pypsa_eur"),
-        length_factor=config_provider("lines", "length_factor"),
+        length_factor=config_provider(
+            "transmission", "electricity", "lines", "length_factor"
+        ),
         planning_horizons=config_provider("scenario", "planning_horizons"),
         countries=config_provider("countries"),
         adjustments=config_provider("adjustments", "sector"),
