+++ scripts/lib/validation/config/transmission.py
@@ -12,6 +12,172 @@ from typing import Literal
 
 from pydantic import BaseModel, Field
 
+from scripts.lib.validation.config._base import ConfigModel
+
+
+class _DynamicLineRatingConfig(ConfigModel):
+    """Configuration for `lines.dynamic_line_rating` settings."""
+
+    activate: bool = Field(
+        False,
+        description="Whether to take dynamic line rating into account.",
+    )
+    cutout: str | list[str] = Field(
+        "default",
+        description="Specifies the weather data cutout file(s) to use.",
+    )
+    correction_factor: float = Field(
+        0.95,
+        description="Factor to compensate for overestimation of wind speeds in hourly averaged wind data.",
+    )
+    max_voltage_difference: float | Literal[False] = Field(
+        False,
+        description="Maximum voltage angle difference in degrees or 'false' to disable.",
+    )
+    max_line_rating: float | Literal[False] = Field(
+        False,
+        description="Maximum line rating relative to nominal capacity without DLR, e.g. 1.3 or 'false' to disable.",
+    )
+
+
+class LinesConfig(BaseModel):
+    """Configuration for `lines` settings."""
+
+    types: dict[float, str] = Field(
+        default_factory=lambda: {
+            63.0: "94-AL1/15-ST1A 20.0",
+            66.0: "94-AL1/15-ST1A 20.0",
+            90.0: "184-AL1/30-ST1A 110.0",
+            110.0: "184-AL1/30-ST1A 110.0",
+            132.0: "243-AL1/39-ST1A 110.0",
+            150.0: "243-AL1/39-ST1A 110.0",
+            220.0: "Al/St 240/40 2-bundle 220.0",
+            300.0: "Al/St 240/40 3-bundle 300.0",
+            330.0: "Al/St 240/40 3-bundle 300.0",
+            380.0: "Al/St 240/40 4-bundle 380.0",
+            400.0: "Al/St 240/40 4-bundle 380.0",
+            500.0: "Al/St 240/40 4-bundle 380.0",
+            750.0: "Al/St 560/50 4-bundle 750.0",
+        },
+        description="Specifies line types to assume for the different voltage levels of the ENTSO-E grid extraction. Should normally handle voltage levels 220, 300, and 380 kV.",
+    )
+    s_max_pu: float = Field(
+        0.7,
+        description="Correction factor for line capacities (`s_nom`) to approximate N-1 security and reserve capacity for reactive power flows.",
+    )
+    s_nom_max: float = Field(
+        float("inf"),
+        description="Global upper limit for the maximum capacity of each extendable line (MW).",
+    )
+    max_extension: float = Field(
+        20000,
+        description="Upper limit for the extended capacity of each extendable line (MW).",
+    )
+    length_factor: float = Field(
+        1.25,
+        description="Correction factor to account for the fact that buses are *not* connected by lines through air-line distance.",
+    )
+    reconnect_crimea: bool = Field(
+        True,
+        description="Whether to reconnect Crimea to the Ukrainian grid.",
+    )
+    under_construction: Literal["zero", "remove", "keep"] = Field(
+        "keep",
+        description="Specifies how to handle lines which are currently under construction.",
+    )
+    dynamic_line_rating: _DynamicLineRatingConfig = Field(
+        default_factory=_DynamicLineRatingConfig,
+        description="Configuration for dynamic line rating.",
+    )
+
+
+class LinksConfig(ConfigModel):
+    """Configuration for `links` settings."""
+
+    p_max_pu: float = Field(
+        1.0,
+        description="Correction factor for link capacities `p_nom`.",
+    )
+    p_min_pu: float = Field(
+        -1.0,
+        description="Correction factor for link capacities `p_nom`.",
+    )
+    p_nom_max: float = Field(
+        float("inf"),
+        description="Global upper limit for the maximum capacity of each extendable DC link (MW).",
+    )
+    max_extension: float = Field(
+        30000,
+        description="Upper limit for the extended capacity of each extendable DC link (MW).",
+    )
+    length_factor: float = Field(
+        1.25,
+        description="Correction factor to account for the fact that buses are *not* connected by links through air-line distance.",
+    )
+    under_construction: Literal["zero", "remove", "keep"] = Field(
+        "keep",
+        description="Specifies how to handle lines which are currently under construction.",
+    )
+
+
+class TransformersConfig(ConfigModel):
+    """Configuration for `transformers` settings."""
+
+    x: float = Field(
+        0.1,
+        description="Series reactance in per unit (p.u.), using `s_nom` as base power of the transformer. Overwritten if `type` is specified.",
+    )
+    s_nom: float = Field(
+        2000.0,
+        description="Limit of apparent power which can pass through branch (MVA). Overwritten if `type` is specified.",
+    )
+    type: str = Field(
+        "",
+        description="Specifies transformer types to assume for the transformers of the ENTSO-E grid extraction.",
+    )
+
+
+class _IncludeConfig(ConfigModel):
+    """Configuration for `electricity.projects.include` settings."""
+
+    tyndp2020: bool = Field(
+        True,
+        description="Whether to integrate the TYNDP 2020 dataset.",
+    )
+    nep: bool = Field(
+        True,
+        description="Whether to integrate the German network development plan dataset.",
+    )
+    manual: bool = Field(
+        True,
+        description="Whether to integrate the manually added transmission projects. They are taken from the previously existing links_tyndp.csv file.",
+    )
+
+
+class ElectricityProjectsConfig(BaseModel):
+    """Configuration for `electricity.projects` settings."""
+
+    enable: bool = Field(
+        True,
+        description="Whether to integrate this transmission projects or not.",
+    )
+    include: _IncludeConfig = Field(
+        default_factory=_IncludeConfig,
+        description="Name of the transmission projects. They should be unique and have to be provided in the `data/transmission_projects` folder.",
+    )
+    skip: list[str] = Field(
+        default_factory=lambda: ["upgraded_lines", "upgraded_links"],
+        description="Type of lines to skip from all transmission projects. Possible values are: `upgraded_lines`, `upgraded_links`, `new_lines`, `new_links`.",
+    )
+    status: list[str] | dict[str, list[str]] = Field(
+        default_factory=lambda: ["under_construction", "in_permitting", "confirmed"],
+        description="Status to include into the model as list or as dict with name of project and status to include. Possible values for status are `under_construction`, `in_permitting`, `confirmed`, `planned_not_yet_permitted`, `under_consideration`.",
+    )
+    new_link_capacity: Literal["zero", "keep"] = Field(
+        "zero",
+        description="Whether to set the new link capacity to the provided capacity or set it to zero.",
+    )
+
 
 class _GabrielFilterConfig(BaseModel):
     """Configuration for `transmission.<carrier>.gabriel_filter` settings."""
@@ -64,6 +230,22 @@ class _TransmissionCarrierConfigElectricity(BaseModel):
     transmission_limit: str = Field(
         "vopt",
         description="Limit on transmission expansion. The first part can be `v` (for setting a limit on line volume) or `c` (for setting a limit on line cost). The second part can be `opt` or a float bigger than one (e.g. 1.25). If `opt` is chosen line expansion is optimised according to its capital cost (where the choice `v` only considers overhead costs for HVDC transmission lines, while `c` uses more accurate costs distinguishing between overhead and underwater sections and including inverter pairs). The setting `v1.25` will limit the total volume of line expansion to 25% of currently installed capacities weighted by individual line lengths. The setting `c1.25` will allow to build a transmission network that costs no more than 25 % more than the current system.",
+    )
+    lines: LinesConfig = Field(
+        default_factory=LinesConfig,
+        description="Transmission lines configuration.",
+    )
+    links: LinksConfig = Field(
+        default_factory=LinksConfig,
+        description="HVDC links configuration.",
+    )
+    transformers: TransformersConfig = Field(
+        default_factory=TransformersConfig,
+        description="Transformers configuration.",
+    )
+    projects: ElectricityProjectsConfig = Field(
+        default_factory=ElectricityProjectsConfig,
+        description="Electricity transmission projects configuration.",
     )
